+++ packages/brick_sqlite_generators/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+* Do not create joins tables for ignored fields
 * Type arguments are stripped from fields when building the `fieldsToSqliteColumns` definition (#31)
 * Fixes a bug where sets were not serialized by SQLite
 * Single siblings are upserted to SQLite as they're received; previously only iterable siblings were upserted